--- .bowerrc
@@ -1,3 +1,0 @@
-{
-  "directory": "app/bower_components"
-}
