--- .yo-rc.json
@@ -1,6 +1,0 @@
-{
-  "generator-mocha": {
-    "ui": "bdd",
-    "rjs": false
-  }
-}
