--- app/manifest.json
@@ -1,62 +1,0 @@
-{
-  "name": "Real Reality",
-  "version": "0.0.11",
-  "manifest_version": 2,
-  "description": "Adds little side panel to some real estate sites which shows you quality of location eg. commutation times, air pollution.",
-  "icons": {
-    "16": "images/icon-16.png",
-    "128": "images/icon-128.png"
-  },
-  "default_locale": "en",
-  "permissions": [
-    "storage",
-    "tabs",
-    "http://*/*",
-    "https://*/*",
-    "pageCapture",
-    "webRequest",
-    "*://*.google.com/",
-    "*://*.googleapis.com/"
-  ],
-  "web_accessible_resources": [
-   "panel.html",
-   "styles/panel.css",
-   "styles/cssnormalize-context-min.css",
-   "bower_components/geocomplete/jquery.geocomplete.min.js",
-   "bower_components/font-awesome/css/font-awesome.min.css",
-   "bower_components/font-awesome/fonts/fontawesome-webfont.eot",
-   "bower_components/font-awesome/fonts/fontawesome-webfont.svg",
-   "bower_components/font-awesome/fonts/fontawesome-webfont.ttf",
-   "bower_components/font-awesome/fonts/fontawesome-webfont.woff",
-   "bower_components/font-awesome/fonts/fontawesome-webfont.woff2"
-  ],
-  "content_scripts": [
-    {
-      "matches": [
-        "*://*.bezrealitky.cz/*",
-        "*://*.sreality.cz/*",
-        "*://*.maxirealitypraha.cz/*"
-      ],
-      "js": [
-        "bower_components/jquery/dist/jquery.min.js",
-        "bower_components/geocomplete/jquery.geocomplete.min.js",
-        "bower_components/vue/dist/vue.min.js",
-        "bower_components/vue-i18n/dist/vue-i18n.min.js",
-        "scripts/contentscript.js"
-      ],
-      "run_at": "document_end",
-      "all_frames": false
-    }
-  ],
-  "content_security_policy": "script-src 'self' https://maps.googleapis.com; object-src 'self'",
-  "browser_action": {
-    "default_icon": {
-      "19": "images/icon-19-off.png",
-      "38": "images/icon-38-off.png"
-    },
-    "default_title": "Real Reality"
-  },
-  "background": {
-    "scripts": ["scripts/background.js"]
-  }
-}
