--- app/styles/main.css
@@ -1,3 +1,0 @@
-body {
-  padding: 20px;
-}
