--- bower.json
@@ -1,15 +1,0 @@
-{
-  "name": "reality",
-  "private": true,
-  "version": "0.0.0",
-  "dependencies": {
-    "font-awesome": "^4.6.3",
-    "jquery": "^3.1.1",
-    "vue": "^2.0.2",
-    "vue-i18n": "^4.6.0",
-    "geocomplete": "^1.7.0"
-  },
-  "devDependencies": {
-    "chai": "^3.5.0"
-  }
-}
