+++ package.json
@@ -1,47 +1,43 @@
 {
   "name": "reality",
+  "version": "0.12.0",
+  "description": "Adds little side panel to some real estate sites which shows you quality of location eg. commutation times, air pollution.",
   "private": true,
-  "engines": {
-    "node": ">=0.8.0"
-  },
   "scripts": {
-    "lint": "eslint -c .eslintrc app/scripts.babel",
-    "test": "mocha --compilers js:babel-core/register test/**/*.spec.js"
-  },
-  "devDependencies": {
-    "babel-core": "^6.7.2",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-es2015-rollup": "^1.2.0",
-    "babel-register": "^6.18.0",
-    "chai": "^3.5.0",
-    "del": "^2.2.0",
-    "eslint": "^3.8.1",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^6.1.2",
-    "gulp-cache": "^0.4.3",
-    "gulp-chrome-manifest": "0.0.13",
-    "gulp-clean-css": "^2.0.3",
-    "gulp-debug": "^2.1.2",
-    "gulp-eslint": "^2.0.0",
-    "gulp-htmlmin": "^1.3.0",
-    "gulp-if": "^2.0.0",
-    "gulp-imagemin": "^2.4.0",
-    "gulp-livereload": "^3.8.1",
-    "gulp-load-plugins": "^1.2.0",
-    "gulp-rollup": "^2.5.1",
-    "gulp-size": "^2.1.0",
-    "gulp-sourcemaps": "^2.2.0",
-    "gulp-uglify": "^1.5.3",
-    "gulp-useref": "^3.0.8",
-    "gulp-zip": "^3.2.0",
-    "jsdom": "^9.8.3",
-    "main-bower-files": "^2.11.1",
-    "mocha": "^3.1.2",
-    "rollup-plugin-babel": "^2.6.1",
-    "run-sequence": "^1.1.5",
-    "wiredep": "^4.0.0"
+    "build": "node scripts/build.js",
+    "lint": "eslint -c .eslintrc ./src/",
+    "start": "node scripts/webserver.js",
+    "test": "jest",
+    "test:ci": "yarn lint && jest --coverage"
   },
   "dependencies": {
-    "gulp-sass": "^2.3.2"
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.1.1",
+    "vue": "^2.0.5",
+    "vue-i18n": "^4.7.1"
+  },
+  "devDependencies": {
+    "babel-core": "^6.18.2",
+    "babel-loader": "^6.2.7",
+    "babel-preset-es2015": "^6.18.0",
+    "chai": "^3.5.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.26.0",
+    "eslint": "^3.10.0",
+    "expose-loader": "^0.7.1",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "fs-extra": "^1.0.0",
+    "html-loader": "^0.4.4",
+    "jest": "^17.0.3",
+    "postcss-loader": "^1.1.1",
+    "precss": "^1.4.0",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.3",
+    "webpack-dev-server": "^1.16.2",
+    "write-file-webpack-plugin": "^3.4.2"
+  },
+  "jest": {
+    "testRegex": "<rootDir>/src/.*\\.spec\\.(js)$",
+    "verbose": true
   }
 }
