+++ src/css/panel.scss
@@ -100,7 +100,7 @@
     color: black;
     margin-left: 15px;
     margin-top: 20px;
-    margin-bottom: 20px;    
+    margin-bottom: 20px;
     width: 80%;
     font-size: 12px;
     padding-left: 5px;
