+++ src/images/icon-128.png
