+++ src/js/i18n/locales.js
@@ -2,8 +2,9 @@ import cs from './cs.js';
 import en from './en.js';
 
 const RRLocales = {
-  cs : cs,
-  en : en
+  cs: cs,
+  en: en,
+  'en-US': en
 };
 
 export default RRLocales;
