+++ src/js/sites/index.js
@@ -1,5 +1,3 @@
-'use strict';
-
 const textOrNull = textElement => {
   if (textElement === null) {
     return null;
