--- test/spec/test.js
@@ -1,11 +1,0 @@
-(function () {
-  'use strict';
-
-  describe('Give it some context', function () {
-    describe('maybe a bit more context here', function () {
-      it('should run here few assertions', function () {
-
-      });
-    });
-  });
-})();
